Checkpoint + id should not be a field + LocalStore -> Store + Remove BaseStore
